rename js -> src
