+++ src/app/proposals-styles.ts
