+++ src/libraries/tequilapi/enums.ts
